533 Move zoning resolution optionsets into their own file
