+++ client/app/optionsets/affected-zoning-resolution.js
@@ -1,0 +1,142 @@
+export const AFFECTED_ZONING_RESOLUTION_ACTION_OPTIONSET = {
+  ZA: {
+    code: 717170000,
+    label: 'ZA',
+  },
+  ZC: {
+    code: 717170001,
+    label: 'ZC',
+  },
+  ZS: {
+    code: 717170002,
+    label: 'ZS',
+  },
+  ZR: {
+    code: 717170003,
+    label: 'ZR',
+  },
+  ZM: {
+    code: 717170008,
+    label: 'ZM',
+  },
+  SD: {
+    code: 717170009,
+    label: 'SD',
+  },
+  SC: {
+    code: 717170010,
+    label: 'SC',
+  },
+  SA: {
+    code: 717170011,
+    label: 'SA',
+  },
+  RS: {
+    code: 717170012,
+    label: 'RS',
+  },
+  RC: {
+    code: 717170013,
+    label: 'RC',
+  },
+  RA: {
+    code: 717170014,
+    label: 'RA',
+  },
+  PS: {
+    code: 717170015,
+    label: 'PS',
+  },
+  PQ: {
+    code: 717170016,
+    label: 'PQ',
+  },
+  PP: {
+    code: 717170017,
+    label: 'PP',
+  },
+  PE: {
+    code: 717170018,
+    label: 'PE',
+  },
+  PC: {
+    code: 717170019,
+    label: 'PC',
+  },
+  NP: {
+    code: 717170020,
+    label: 'NP',
+  },
+  MY: {
+    code: 717170021,
+    label: 'MY',
+  },
+  MM: {
+    code: 717170022,
+    label: 'MM',
+  },
+  ML: {
+    code: 717170023,
+    label: 'ML',
+  },
+  ME: {
+    code: 717170024,
+    label: 'ME',
+  },
+  MC: {
+    code: 717170025,
+    label: 'MC',
+  },
+  LD: {
+    code: 717170026,
+    label: 'LD',
+  },
+  HU: {
+    code: 717170027,
+    label: 'HU',
+  },
+  HP: {
+    code: 717170028,
+    label: 'HP',
+  },
+  HO: {
+    code: 717170029,
+    label: 'HO',
+  },
+  HN: {
+    code: 717170030,
+    label: 'HN',
+  },
+  HG: {
+    code: 717170031,
+    label: 'HG',
+  },
+  HD: {
+    code: 717170032,
+    label: 'HD',
+  },
+  HC: {
+    code: 717170033,
+    label: 'HC',
+  },
+  HA: {
+    code: 717170034,
+    label: 'HA',
+  },
+  GF: {
+    code: 717170035,
+    label: 'GF',
+  },
+  CM: {
+    code: 717170036,
+    label: 'CM',
+  },
+  BF: {
+    code: 717170037,
+    label: 'BF',
+  },
+  BD: {
+    code: 717170038,
+    label: 'BD',
+  },
+};
